Folded JCuda{Context,Object} to GPU{Context,Object}
